+++ crates/nil/src/vfs.rs
@@ -92,7 +92,7 @@ impl Vfs {
                 return Err(format!("Invalid range {:?}", del_range).into());
             }
             let mut buf = String::with_capacity(
-                text.len() - usize::from(del_range.len()) + ins_text.len() as usize,
+                text.len() - usize::from(del_range.len()) + ins_text.len(),
             );
             buf += &text[..usize::from(del_range.start())];
             buf += ins_text;
